+++ service/database/db-structs.go
@@ -9,8 +9,8 @@ type DatabasePhoto struct {
 	Id uint64 `json:"id"`
 	Url string `json:"url"`
 	Date string `json:"date"`
-	LikeCount uint64 `json:"like_count"`
-	CommentCount uint64 `json:"comment_count"`
+	LikeCount int `json:"like_count"`
+	CommentCount int `json:"comment_count"`
 }
 
 type DatabaseComment struct {
@@ -20,9 +20,9 @@ type DatabaseComment struct {
 
 type DatabaseProfile struct {
 	User DatabaseUser `json:"user"`
-	PhotosCount uint64 `json:"photos_count"`
-	FollowersCount uint64 `json:"followers_count"`
-	FollowingCount uint64 `json:"following_count"`
+	PhotosCount int `json:"photos_count"`
+	FollowersCount int `json:"followers_count"`
+	FollowingCount int `json:"following_count"`
 }
 
 type DatabaseStream struct {
